Improving CSS and adding static files
Learned how to use some flexbox and some smart solutions.
What I did in the index.html and style.css was try to center a flexbox element while the other is aligned to the right, and the idea is to create a flex for EACH item while they are all inside a master one. Pretty cool